241103 양혁수 자유게시판 댓글 부분 수정중

+++ dr/src/main/java/com/dr/dto/board/FreeBoardCommentDTO.java
@@ -15,8 +15,8 @@ public class FreeBoardCommentDTO {
     private String replyWriteDate;
     private String replyModifyDate;
     private String photoLocal;
-    private int userNumber;
-    private int boardNumber;
-    private int recipeNumber;
+    private Long userNumber;
+    private Long boardNumber;
+    private Long recipeNumber;
     private String userNickName;
 }
